Humanize scanner test strategy
This commit improves tests for slide's scanner and tokens by converting
scanned tokens to a human-readable string form, and then comparing that
string form to a string expectation. With a macro that abstracts
creation of a scanner and translation of the program input to this
human-readable string, this test strategy should make it very easy to
see what tests there currently are, and add/modify more as necessary.

Closes #4.