Apply rules on associative nodes
We do this by flattening binary expressions, rewriting with various
associativities, and running the rewritten expressions on various rules.
If a rewritten expression can be applied by the rule, we use it.
Otherwise we use the rewritten expression whose partial transformation
is the "least complex" (i.e. has the least # of operations).

This still doesn't fix something like `"1 + 2 -b + 3 + a"`. I need to
think more on that. @luke-bhan if you'd like, have a go on it.

Closes #55