Small refactor of evaluation process
I am looking into some way to evaluate expressions like

```
1 + 2 - b + 3 + a
```

To do this, I think the best way forward is to flatten math expressions
with negation rewritten as subtraction. That is, the expression above
should become

```
1 + 2 + -b + 3 + a
```

The code to do this should belong to the `add` rule. As a precursor,
this commit sets up `add` to fold constants in a flattened expression,
and changes symmetric expression pattern matching to only happen in a
string matching context (that is the only place we really care about it,
since a function rule should be able to handle any analysis it needs
itself).