Flatten expressions commutatively
This commit adds an option to flatten/fold expressions a little bit smarter
than we currently do by flattening and folding an expression
commutatively.

Not sure if there should be a better name, like `flat_fold` or similar?
Or we can rename the current usages of `flatten` and `unflatten`, which
don't do any expression folding, to `rollup` and `unroll` or similar.
Feedback wanted on this.

Closes #92