Flatten multiplication and division
Also makes some chnages to how we format emitted expressions, since
previously we did it incorrectly. There's a huge amount of documentation
in this PR... I hope that this is a blessing :)

The absolute biggest bet is that this change succeeds in lowering the
expression we present in our readme! So, I think we are ready to cut the
0.1 release after this fix. cc @luke-bhan @timothycobrien

Closes #106