Add algorithm for converting decimal to fraction
This commit adds an algorithm for converting a decimal number to its
fraction representation using a convergent algorithm that walks the
[Stern–Brocot tree](https://en.wikipedia.org/wiki/Stern%E2%80%93Brocot_tree).

If the algorithm fails to find an exact fractional representation after
a specified number of iterations, it bails out. This is to allow the
user to decide at which point they want to stop searching for a
fraction.