Add format emit strategies to libslide
Currently, we determine on how a slide invocation should be emitted on
the application side, routing to `fmt::Display`, `fmt::Debug`, or
`Grammar#s_expr` accordingly. This won't scale very well, and in truth
our emission strategies should be the basis for other core rust traits
(like the `fmt` ones) implemented for the libslide IR.

This commit adds an `emit` module that provides an `EmitFormat` enum
replacing the existing `OutputForm` enum on the application side, and
exposes an `Emit` trait used to describe how a type can be emitted.
Eventually all types implementing `Emit` should implement all of its
member functions, but to ease migration, by default only `emit_pretty`
must be implemented.

The nice thing about this structure is that `fmt::Display` comes for
free if your type implements `emit_pretty`, and `emit_debug` comes for
free if your type implements `fmt::Debug`. Eventually, I would like to
reconcile the coupling between the `fmt` macros and `Emit` entirely,
maybe using a proc macro, but that is outside the scope of this change
and requires some forethought.

This work prepares for a LaTeX emit strategy (also low priority, but if
a new contributor would like to add it, this commit lays the groundwork
to make it easier).

Part of #162