Slide web UI
This commit adds a web UI for slide, continuing off @lukebhan's work in
 #156. The UI is static, using a wasm build of slide to run in the
browser. The UI will be built continuously to http://yslide.github.io on
each commit to master.

To enable easier iteration during development, `ladder www` has been
added as a command to spin up a local file server for the UI and watch
changes to slide. In practice, this should be useful for development
too.

The UI supports changing all options slide currently supports and
preserves state in the URL, enabling easy sharing.

In later iterations we can add information on current commit/version,
links to a bug report, etc.

I highly recommend checking this branch out locally to review the UI. A
screenshot is attached in the PR associated with this commit.

Closes #115
Closes #169