Add interner
This commit adds an interner for slide Grammars. See
libslide/src/grammar/intern.rs for more details. The purpose of this is
to reduce duplication of grammar instances that already exist by
aggregating all instances in a common pool, and returning an opaque
reference to the instance rather than the instance directly. Operations
on the instance can still be done through the intern reference via
dereferences or methods on traits common to Grammars.

The intern strategy is

```
Grammar -> String Hash -> 32-bit identifier
^^^^^^^                   ^^^^^^^^^^^^^^^^^
instance                 interned reference
```

The 32-bit identifier is then mapped back to a Grammar when we need to
dereference it.

Closes #101