Add support for spans on expressions
This prepares for a "linter" pass that will walk a parse tree and detect
"bad" patterns, like `++++5` which is trivially reducible to `5`.

Future work may also use these spans to provide intellegent suggestions
on reducibility, for example noting that `3 + 4` is the span that was
converted to `7` in an expression like `2 * 3 / (3 + 4)`.

Also, this change has nearly convinced me that having the interner in a
state as we do now is a useless and bad idea, since it's (1) already
pretty complicated and (2) making expressions unique by their span
defeats the usefulness of caching structurally identical expressions.

Part of #135