Record leading trivia as full span on tokens
This will be used to distinguish newline-separated statements in a
future patch adding support for statement delimination across multiple
lines. That is, we want

```
1 + 2
2 + 3
```

to parse as two expression statements, but

```
1 + 2 2 + 3
```

not to do so.

To support this, we need information about whether the start of an
expression is preceded by a newline, which leading trivia information
will provide.