Support autofixes on diagnostics
This patch adds support for programatic declaration of autofixes that
can be applied to diagnostics, and their implementation for many of the
current diagnostics. Presently these autofixes are just formatted and
displayed during emit, but later on they can be used to provide an
`--autofix` or similar CLI flag that will automatically apply fixes to a
program, or for use in a language service (cf #303).

Closes #304
Milestone 0.0.2 (second to last one!)
PTAL @lukebhan