Server document model
This model will be used to parse and analyze multiple slide programs
within a single document. The idea is that the user can supply a
"document parser" that will discover blocks of slide programs in a
document type the user wants to use the server for. This patch already
implements the `DocumentParser` type for this use case, discovery of
document parsers from the server initialization options, and validation
of the parsers.

There's some other cleanup here too because I'm bad at concentrating.

cf #290