Program, Document, and Document Registry models
A Program is a single instance of a slide program. A Program contains
some Analysis work which will be used to answer server queries. A
Document corresponds to a text document for which the server handles
queries, and can contain multiple Programs. A Document Registry is a
stateful database of Documents known to a server session and handles the
application of Document changes, as observed by the server session.